+++ weblate/trans/tests/test_acl.py
@@ -23,6 +23,7 @@
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User, Group
 
+from weblate.billing.models import Plan, Billing
 from weblate.trans.models import Project
 from weblate.trans.tests.test_views import FixtureTestCase
 
@@ -226,3 +227,28 @@ class ACLViewTest(FixtureTestCase):
             follow=True
         )
         self.assertContains(response, 'No matching user found!')
+
+    def test_change_access(self):
+        self.add_acl()
+        url = reverse('change-access', kwargs=self.kw_project)
+        self.project.add_user(self.user, '@Administration')
+
+        # No permissions
+        response = self.client.post(url, {'access_control': 0})
+        self.assertEqual(response.status_code, 403)
+
+        # Allow editing by creating billing plan
+        plan = Plan.objects.create()
+        billing = Billing.objects.create(plan=plan, user=self.user)
+        billing.projects.add(self.project)
+
+        # Editing should now work
+        response = self.client.post(
+            url,
+            {'access_control': Project.ACCESS_PROTECTED}
+        )
+        self.assertRedirects(response, self.access_url)
+
+        # Verify change has been done
+        project = Project.objects.get(pk=self.project.pk)
+        self.assertEqual(project.access_control, Project.ACCESS_PROTECTED)