+++ docs/changes.rst
@@ -16,6 +16,7 @@ Released on ? 2017.
 * More fine grained per project ACL configuration.
 * Default value of BASE_DIR has been changed.
 * Added two step account removal to prevent accidental removal.
+* Project access control settings is now editable.
 
 weblate 2.16
 ------------